Update device_tracker.py
So after some testing, it appears that the problem is not with the data being provided, but more that HA seems to mark a device tracer as unavailable unless the timestamp is different?  Anyways, probably a nicer way to get it done, but this at least is working for me in that when the API succeeds, I get an update, and when not, it at least holds its current state.